+++ src/com/epm/gdsa/point/PointService.java
@@ -12,6 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.epm.gdsa.project.Project;
 import com.epm.gdsa.user.User;
+import com.epm.utils.PublicUtils;
 
 @Transactional
 @Component(value="pointService")
@@ -143,26 +144,44 @@ public class PointService {
 	
 
 	public Map<String, Object> getPointByIds(String keys,Integer page,Integer pageSize,Integer[] ids) {
-		System.out.println("this 0");
 		DetachedCriteria criteria = pointDao.getCriteriaByIds(ids);
-		System.out.println("this 1");
 		Map<String, Object> map = getMapByKeysAndPage(keys,page,pageSize,criteria);
-		System.out.println(map.size());
 		return map;
 	}
 
 	public Map<String, Object> updateByIds(String keys, Integer[] idsIntegers,
-			Point point, User user2) {
+			Point point, User loginUser) {
 		Map<String, Object> map = new HashMap<String, Object>();
 		Map<String, Object> theMap = null;
-		map.put("code", 200);
+		map.put("code", 201);
 		for (Integer integer : idsIntegers) {
 			Point point2 = pointDao.getById(integer);
-			point2 = getNewPointByKeys(point2,point,keys);
 			if(point2 != null){
-				pointDao.update(point2);
+				Project theProject = point2.getProject();
+				if(theProject == null){
+					map.put("code", 404);
+					if(theMap == null){
+						theMap = new HashMap<String, Object>();
+						theMap.put("error", "id为"+integer+"的数据修改失败:项目不存在;");
+					}else{
+						theMap.put("error",theMap.get("error")+"id为"+integer+"的数据修改失败:项目不存在;");
+					}
+				}else if(!theProject.getUser().getUserId().equals(loginUser.getUserId()) && !point2.getUser().getUserId().equals(loginUser.getUserId())){
+					map.put("code", 401);
+					if(theMap == null){
+						theMap = new HashMap<String, Object>();
+						theMap.put("error", "id为"+integer+"的数据修改失败:您不具有权限;");
+					}else{
+						theMap.put("error",theMap.get("error")+"id为"+integer+"的数据修改失败:您不具有权限;");
+					}
+				}else{
+					//具有权限时的业务逻辑
+					point2 = getNewPointByKeys(point2,point,keys);
+					pointDao.update(point2);
+				}
+				
 			}else{
-				map.put("code", 400);
+				map.put("code", 404);
 				if(theMap == null){
 					theMap = new HashMap<String, Object>();
 					theMap.put("error", "id为"+integer+"的数据修改失败;");
@@ -176,16 +195,36 @@ public class PointService {
 		return map;
 	}
 
-	public Map<String, Object> deleteByIds(Integer[] idsIntegers, User user2) {
+	public Map<String, Object> deleteByIds(Integer[] idsIntegers, User loginUser) {
 		Map<String, Object> map = new HashMap<String, Object>();
 		Map<String, Object> theMap = null;
-		map.put("code", 200);
+		map.put("code", 201);
 		for (Integer integer : idsIntegers) {
 			Point point2 = pointDao.getById(integer);
 			if(point2 != null){
-				pointDao.delete(point2);
+				Project theProject = point2.getProject();
+				if(theProject == null){
+					map.put("code", 404);
+					if(theMap == null){
+						theMap = new HashMap<String, Object>();
+						theMap.put("error", "id为"+integer+"的数据修改失败:项目不存在;");
+					}else{
+						theMap.put("error",theMap.get("error")+"id为"+integer+"的数据修改失败:项目不存在;");
+					}
+				}else if(!theProject.getUser().getUserId().equals(loginUser.getUserId()) && !point2.getUser().getUserId().equals(loginUser.getUserId())){
+					map.put("code", 401);
+					if(theMap == null){
+						theMap = new HashMap<String, Object>();
+						theMap.put("error", "id为"+integer+"的数据修改失败:您不具有权限;");
+					}else{
+						theMap.put("error",theMap.get("error")+"id为"+integer+"的数据修改失败:您不具有权限;");
+					}
+				}else{
+					//具有权限时的业务逻辑
+					pointDao.delete(point2);
+				}
 			}else{
-				map.put("code", 400);
+				map.put("code", 404);
 				if(theMap == null){
 					theMap = new HashMap<String, Object>();
 					theMap.put("error", "id为"+integer+"的数据删除失败：数据不存在;");
