+++ src/com/epm/gdsa/project/ProjectService.java
@@ -179,17 +179,28 @@ public class ProjectService {
 	}
 
 	public Map<String, Object> updateByIds(String keys, Integer[] idsIntegers,
-			Project project, User user2) {
+			Project project, User loginUser) {
 		Map<String, Object> map = new HashMap<String, Object>();
 		Map<String, Object> theMap = null;
-		map.put("code", 200);
+		map.put("code", 201);
 		for (Integer integer : idsIntegers) {
 			Project project2 = projectDao.getById(integer);
-			project2 = getNewProjectByKeys(project2,project,keys);
 			if(project2 != null){
-				projectDao.update(project2);
+				if(project2.getUser().getUserId().equals(loginUser.getUserId())){
+					project2 = getNewProjectByKeys(project2,project,keys);
+					projectDao.update(project2);
+				}else{
+					map.put("code", 401);
+					if(theMap == null){
+						theMap = new HashMap<String, Object>();
+						theMap.put("error", "id为"+integer+"的数据修改失败:您不具有权限;");
+					}else{
+						theMap.put("error",theMap.get("error")+"id为"+integer+"的数据修改失败:您不具有权限;");
+					}
+				}
+				
 			}else{
-				map.put("code", 400);
+				map.put("code", 404);
 				if(theMap == null){
 					theMap = new HashMap<String, Object>();
 					theMap.put("error", "id为"+integer+"的数据修改失败;");
@@ -203,21 +214,31 @@ public class ProjectService {
 		return map;
 	}
 
-	public Map<String, Object> deleteByIds(Integer[] idsIntegers, User user2) {
+	public Map<String, Object> deleteByIds(Integer[] idsIntegers, User loginUser) {
 		Map<String, Object> map = new HashMap<String, Object>();
 		Map<String, Object> theMap = null;
-		map.put("code", 200);
+		map.put("code", 204);
 		for (Integer integer : idsIntegers) {
 			Project project2 = projectDao.getById(integer);
 			if(project2 != null){
-				projectDao.delete(project2);
+				if(project2.getUser().getUserId().equals(loginUser.getUserId())){
+					projectDao.delete(project2);
+				}else{
+					map.put("code", 401);
+					if(theMap == null){
+						theMap = new HashMap<String, Object>();
+						theMap.put("error", "id为"+integer+"的数据删除失败:您不具有权限;");
+					}else{
+						theMap.put("error",theMap.get("error")+"id为"+integer+"的数据删除失败:您不具有权限;");
+					}
+				}
 			}else{
-				map.put("code", 400);
+				map.put("code", 404);
 				if(theMap == null){
 					theMap = new HashMap<String, Object>();
 					theMap.put("error", "id为"+integer+"的数据删除失败：数据不存在;");
 				}else{
-					theMap.put("error",theMap.get("error")+"id为"+integer+"的数据修改失败：数据不存在;");
+					theMap.put("error",theMap.get("error")+"id为"+integer+"的数据删除失败：数据不存在;");
 				}
 			}
 		}
